+++ src/test/java/bridge/BridgeSizeGetterTest.java
@@ -1,5 +1,6 @@
 package bridge;
 
+import bridge.input.getter.BridgeSizeGetter;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
